+++ previous-versions/files/RELEASES.md
@@ -605,3 +605,13 @@ Based on:
 - [python v0.1.1] previous-versions/files
 ### Releases
 - [PyPI v0.1.1] https://pypi.org/project/codat-files/0.1.1 - previous-versions/files
+
+## 2023-09-05 00:01:06
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.78.3 (2.96.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.2] previous-versions/files
+### Releases
+- [PyPI v0.1.2] https://pypi.org/project/codat-files/0.1.2 - previous-versions/files